feat: connect story api
